style: 메인페이지 best 농장 가로 스크롤 처리
